gateway_service/requirements.txt
fastapi==0.110.0
uvicorn[standard]
python-magic
httpx
pydantic
python-multipart